requirements.txt
black
python-dotenv
slack-bolt
openai
notion-client==2.7.0
md2notionpage
cachetools
requests
notion2md
langchain-community
langchain-openai
langgraph
langchain-google-community
tavily-python
beautifulsoup4
lxml
aiohttp
browser-use
markdown
pypandoc_binary
tabulate[widechars]
redis
PyGithub
pytest
fastapi
uvicorn[standard]
pydantic
boto3
